LayoutTests/storage/indexeddb/objectstore-basics.html

<html>
<head>
<link rel="stylesheet" href="../../fast/js/resources/js-test-style.css">
<script src="../../fast/js/resources/js-test-pre.js"></script>
<script src="../../fast/js/resources/js-test-post-function.js"></script>
<script src="resources/shared.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>

description("Test the basics of IndexedDB's IDBObjectStore.");
if (window.layoutTestController) 
    layoutTestController.waitUntilDone();

function test()
{
    result = evalAndLog("indexedDB.open('name', 'description')");
    verifyResult(result);
    result.onsuccess = openSuccess;
    result.onerror = unexpectedErrorCallback;
}

function openSuccess()
{
    debug("openSuccess():");
    verifySuccessEvent(event);
    window.db = evalAndLog("db = event.result");

    result = evalAndLog("db.setVersion('new version')");
    verifyResult(result);
    result.onsuccess = setVersionSuccess;
    result.onerror = unexpectedErrorCallback;
}

function setVersionSuccess()
{
    debug("setVersionSuccess():");
    verifySuccessEvent(event);
    window.trans = evalAndLog("trans = event.result");
    shouldBeTrue("trans !== null");
    trans.onabort = unexpectedAbortCallback;

    deleteAllObjectStores(db, createObjectStore);
}

function createObjectStore()
{
    debug("creatObjectStore():");
    var store = evalAndLog("store = db.createObjectStore('storeName', null)");
    var storeNames = evalAndLog("storeNames = db.objectStores");

    shouldBeEqualToString("store.name", "storeName");
    shouldBeNull("store.keyPath");
    shouldBe("storeNames.contains('storeName')", "true");
    shouldBe("storeNames.length", "1");
    // FIXME: test all of object store's methods.

    debug("Ask for a store that doesn't exist:");
    try {
        index = evalAndLog("index = store.index('asdf')");
        shouldBeNull("index"); // Returning null is wrong, but less wrong than returning an actual object!
        testPassed("index is null");
        // FIXME: testFailed("Asking for a store that doesn't exist should have thrown.");
    } catch (err) {
        // FIXME: Verify the correct exception thrown.
    }

    createIndex();
}

function createIndex()
{
    debug("createIndex():");
    window.index = evalAndLog("store.createIndex('indexName', 'x', true)"); // true == unique requirement.
    shouldBeTrue("index !== null");
    shouldBeTrue("store.indexNames.contains('indexName')");
    index = evalAndLog("index = store.index('indexName')");
    shouldBeTrue("index !== null");

    debug("Ask for a store that doesn't exist:");
    try {
        index = evalAndLog("index = store.index('asdf')");
        shouldBeNull("index"); // Returning null is wrong, but less wrong than returning an actual object!
        testPassed("index is null.");
        // FIXME: testFailed("Asking for a store that doesn't exist should have thrown.");
    } catch (err) {
        // FIXME: Verify the correct exception thrown.
    }

    result = evalAndLog("store.add({x: 'value'}, 'key')");
    verifyResult(result);
    result.onsuccess = addSuccess;
    result.onerror = unexpectedErrorCallback;
}

function addSuccess()
{
    debug("addSuccess():");
    verifySuccessEvent(event);
    shouldBeEqualToString("event.result", "key");

    result = evalAndLog("event.source.add({x: 'value'}, 'zzz')");
    verifyResult(result);
    result.onsuccess = unexpectedSuccessCallback;
    result.onerror = addAgainFailure;
}

function addAgainFailure()
{
    debug("addAgainFailure():");
    verifyErrorEvent(event);
    // FIXME: This error code needs to be specced.
    shouldBe("event.code", "IDBDatabaseException.UNKNOWN_ERR");

    transaction = evalAndLog("db.transaction()");
    transaction.onabort = unexpectedErrorCallback;
    var store = evalAndLog("store = transaction.objectStore('storeName')");

    result = evalAndLog("store.get('key')");
    verifyResult(result);
    result.onsuccess = getSuccess;
    result.onerror = unexpectedErrorCallback;
}

function getSuccess()
{
    debug("getSuccess():");
    verifySuccessEvent(event);
    shouldBeEqualToString("event.result.x", "value");
    var store = evalAndLog("store = event.source");

    result = evalAndLog("store.remove('key')");
    verifyResult(result);
    result.onsuccess = removeSuccess;
    result.onerror = unexpectedErrorCallback;
}

function removeSuccess()
{
    debug("removeSuccess():");
    verifySuccessEvent(event);
    shouldBeNull("event.result");
    done();
}

test();

var successfullyParsed = true;

</script>
</body>
</html>
